WIP move audit logs frontend to ee
