+++ packages/core/admin/admin/src/hooks/useSettingsMenu/init.js
@@ -25,13 +25,6 @@ const init = (initialState, { settings, shouldUpdateStrapi }) => {
       intlLabel: { id: 'Settings.permissions', defaultMessage: 'Administration Panel' },
       links: [
         {
-          intlLabel: { id: 'global.auditLogs', defaultMessage: 'Audit Logs' },
-          to: '/settings/audit-logs?pageSize=50&page=1&sort=date:DESC',
-          id: 'auditLogs',
-          isDisplayed: false,
-          permissions: adminPermissions.settings.auditLogs.main,
-        },
-        {
           intlLabel: { id: 'global.roles', defaultMessage: 'Roles' },
           to: '/settings/roles',
           id: 'roles',
