+++ packages/core/admin/admin/src/pages/SettingsPage/utils/defaultRoutes.js
@@ -60,17 +60,6 @@ const defaultRoutes = [
   {
     async Component() {
       const component = await import(
-        /* webpackChunkName: "admin-audit-logs" */ '../pages/AuditLogs/ProtectedListPage'
-      );
-
-      return component;
-    },
-    to: '/settings/audit-logs',
-    exact: true,
-  },
-  {
-    async Component() {
-      const component = await import(
         /* webpackChunkName: "webhook-edit-page" */ '../pages/Webhooks/ProtectedCreateView'
       );
