+++ packages/core/admin/ee/admin/hooks/useSettingsMenu/utils/customGlobalLinks.js
@@ -12,6 +12,17 @@ const ssoGlobalRoutes = strapi.features.isEnabled(strapi.features.SSO)
     ]
   : [];
 
-const customGlobalLinks = [...ssoGlobalRoutes];
+const auditLogsRoutes = [
+  // TODO check if feature enabled
+  {
+    intlLabel: { id: 'global.auditLogs', defaultMessage: 'Audit Logs' },
+    to: '/settings/audit-logs?pageSize=50&page=1&sort=date:DESC',
+    id: 'auditLogs',
+    isDisplayed: false,
+    permissions: adminPermissions.settings.auditLogs.main,
+  },
+];
+
+const customGlobalLinks = [...ssoGlobalRoutes, ...auditLogsRoutes];
 
 export default customGlobalLinks;
