+++ packages/core/admin/ee/admin/pages/SettingsPage/pages/AuditLogs/ListView/PaginationFooter/index.js
