+++ packages/core/admin/ee/admin/pages/SettingsPage/pages/AuditLogs/ListView/utils/tableHeaders.js
