+++ packages/core/admin/ee/admin/pages/SettingsPage/pages/AuditLogs/ProtectedListPage/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { CheckPagePermissions } from '@strapi/helper-plugin';
-import adminPermissions from '../../../../../permissions';
+import adminPermissions from '../../../../../../../admin/src/permissions';
 import ListView from '../ListView';
 
 const ProtectedListPage = () => (
