+++ packages/core/admin/ee/admin/pages/SettingsPage/utils/customRoutes.js
@@ -1,4 +1,17 @@
-const routes = [];
+const routes = [
+  // TODO check if feature enabled
+  {
+    async Component() {
+      const component = await import(
+        /* webpackChunkName: "admin-audit-logs" */ '../pages/AuditLogs/ProtectedListPage'
+      );
+
+      return component;
+    },
+    to: '/settings/audit-logs',
+    exact: true,
+  },
+];
 
 if (strapi.features.isEnabled(strapi.features.SSO)) {
   routes.push({
